Updated self-explorer.py
1. Delete self-reflection
2. Editing prompt to add asking question logic, modify model.py accordingly.
3. Some other stuff I forgot =。=
